Introduce TtlCheckInterval structure
